Work aggregation interface refactoring
